add missing files
